requirements.txt
asgiref==3.8.1
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
pytz==2024.1
sqlparse==0.5.1
typing_extensions==4.12.2
